+++ packages/editor/src/RichContentEditor/RichContentEditor.jsx
@@ -115,6 +115,7 @@ class RichContentEditor extends Component {
       t,
       config = {},
       textAlignment,
+      locale,
     } = this.props;
     const { theme } = this.state;
     const buttons = { pluginButtons, pluginTextButtons };
@@ -135,6 +136,7 @@ class RichContentEditor extends Component {
       getToolbarSettings: config.getToolbarSettings,
       uiSettings: config.uiSettings,
       config,
+      locale,
     });
   }